fix: bookmarking on post content
